+++ src/preprocessing/preprocessing_pass.h
@@ -39,6 +39,7 @@
 #include "preprocessing/preprocessing_pass_context.h"
 #include "smt/smt_engine_scope.h"
 #include "smt/term_formula_removal.h"
+#include "theory/logic_info.h"
 #include "theory/substitutions.h"
 
 namespace CVC4 {
